DL/CNN/CarBrand_Prediction/requirement.txt
flask
numpy
pandas
scikit-learn
tensorflow
Werkzeug
pillow
